pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "psy-lab"
version = "0.1.0"
description = "Large-Scale In-Silico Social-Psychology Experiments with LLM Ensembles"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Social Psych Arena Team"}
]
keywords = ["llm", "social-psychology", "experiments", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.15.0",
    "detoxify>=0.5.0",
    "textstat>=0.7.3",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
psy-lab = "psy_lab.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 